Core data and links to match reports for Euro 2016
data.txt contains JSON-formatted core info about the teams, the stages, and the matches; links-1.txt to links-51.txt contain title, description, and link to match reports for each of the matches. The filename of the match report links contains the numeric ID of the match used in data.txt.